requirements.txt
Pillow==8.3.1
numpy==1.21.2
opencv-python==4.5.3.56
tensorflow==2.6.0
